refactor CMake configuration
