--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "cmake"]
-	path = cmake
-	url = https://github.com/VlinderSoftware/cmake.git
