+++ CMakeCPack.cmake
@@ -1,0 +1,50 @@
+set(CPACK_PACKAGE_NAME "${PROJECT_NAME}")
+set(CPACK_PACKAGE_VENDOR "Vlinder Software")
+set(CPACK_PACKAGE_DESCRIPTION_FILE "${PROJECT_SOURCE_DIR}/INSTALL.md")
+set(CPACK_PACKAGE_DESCRIPTION_SUMMARY "exceptions: generic exceptions and contract library")
+set(CPACK_RESOURCE_FILE_LICENSE "${PROJECT_SOURCE_DIR}/LICENSE.md")
+
+set(CPACK_PACKAGING_INSTALL_PREFIX "/usr")
+#set(CPACK_PACKAGING_INSTALL_PREFIX "/opt/${PROJECT_NAME}")
+
+set(CPACK_PACKAGE_VERSION_MAJOR "${PROJECT_VERSION_MAJOR}")
+set(CPACK_PACKAGE_VERSION_MINOR "${PROJECT_VERSION_MINOR}")
+set(CPACK_PACKAGE_VERSION_PATCH "${PROJECT_VERSION_PATCH}")
+
+set(CPACK_SOURCE_IGNORE_FILES "${PROJECT_BINARY_DIR};/.git/;.gitignore")
+set(CPACK_SOURCE_GENERATOR "ZIP;TGZ")
+set(CPACK_GENERATOR "ZIP;TGZ")
+
+if(UNIX)
+  if(CMAKE_SYSTEM_NAME MATCHES Linux)
+    list(APPEND CPACK_GENERATOR "DEB")
+    set(CPACK_DEBIAN_PACKAGE_MAINTAINER "Vlinder Software")
+    set(CPACK_DEBIAN_PACKAGE_SECTION "devel")
+    set(CPACK_DEBIAN_PACKAGE_DEPENDS "")
+
+    list(APPEND CPACK_GENERATOR "RPM")
+    set(CPACK_RPM_PACKAGE_RELEASE "1")
+    set(CPACK_RPM_PACKAGE_LICENSE "Apache2")
+    set(CPACK_RPM_PACKAGE_REQUIRES "")
+  endif()
+endif()
+
+if(WIN32 OR MINGW)
+  list(APPEND CPACK_GENERATOR "NSIS")
+  set(CPACK_NSIS_PACKAGE_NAME "exceptions")
+  set(CPACK_NSIS_CONTACT "Vlinder Software")
+  set(CPACK_NSIS_ENABLE_UNINSTALL_BEFORE_INSTALL ON)
+endif()
+
+if(APPLE)
+  list(APPEND CPACK_GENERATOR "Bundle")
+  set(CPACK_BUNDLE_NAME "exceptions")
+  configure_file(${PROJECT_SOURCE_DIR}/cmake/Info.plist.in Info.plist @ONLY)
+  set(CPACK_BUNDLE_PLIST ${CMAKE_CURRENT_BINARY_DIR}/Info.plist)
+  set(CPACK_BUNDLE_ICON ${PROJECT_SOURCE_DIR}/cmake/coffee.icns)
+endif()
+
+message(STATUS "CPack generators: ${CPACK_GENERATOR}")
+
+include(CPack)
+
