+++ CMakeLists.txt
@@ -1,100 +1,61 @@
-cmake_minimum_required(VERSION 3.7)
-project(exceptions VERSION 1.0.0 LANGUAGES CXX)
+# CMake 3.6 needed for IMPORTED_TARGET option
+# to pkg_search_module
+cmake_minimum_required(VERSION 3.6 FATAL_ERROR)
 
-if (EXISTS ${CMAKE_SOURCE_DIR}/cmake/settings.cmake)
-	include(${CMAKE_SOURCE_DIR}/cmake/settings.cmake)
-else()
-	set(VLINDER_TARGET_PLATFORM_NAME posix)
+project(exceptions
+  LANGUAGES CXX
+  VERSION 1.0.0
+  )
+
+# <<< General set up >>>
+
+set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_EXTENSIONS OFF)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+message(STATUS "Project will be installed to ${CMAKE_INSTALL_PREFIX}")
+
+if(NOT CMAKE_BUILD_TYPE)
+  set(CMAKE_BUILD_TYPE Release CACHE STRING "Build type" FORCE)
 endif()
+message(STATUS "Build type set to ${CMAKE_BUILD_TYPE}")
 
 include(GNUInstallDirs)
 
-if (${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_SOURCE_DIR})
-set(CPACK_PACKAGE_NAME exceptions)
-set(CPACK_PACKAGE_VENDOR "Vlinder Software")
-set(CPACK_PACKAGE_CONTACT "Ronald Landheer-Cieslak <rlc@vlinder.ca>")
-set(CPACK_PACKAGE_VERSION_MAJOR 1)
-set(CPACK_PACKAGE_VERSION_MINOR 0)
-set(CPACK_PACKAGE_VERSION_PATCH 0)
-set(CPACK_PACKAGE_DESCRIPTION "Generic exceptions support library")
-set(CPACK_PACKAGE_DESCRIPTION_SUMMARY "Generic exceptions support library")
-set(CPACK_PACKAGE_DESCRIPTION_FILE ${CMAKE_CURRENT_SOURCE_DIR}/CPack.Description.txt)
-set(CPACK_PACKAGE_HOMEPAGE_URL "http://vlinder.ca")
-set(CPACK_RESOURCE_FILE_README ${CMAKE_CURRENT_SOURCE_DIR}/README.md)
-set(CPACK_RESOURCE_FILE_LICENSE ${CMAKE_CURRENT_SOURCE_DIR}/LICENSE.md)
-set(CPACK_GENERATOR DEB)
-include(CPack)
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY
+	${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY
+	${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY
+	${PROJECT_BINARY_DIR}/${CMAKE_INSTALL_BINDIR})
+
+# Offer the user the choice of overriding the installation directories
+set(INSTALL_LIBDIR ${CMAKE_INSTALL_LIBDIR}
+	CACHE PATH "Installation directory for libraries")
+set(INSTALL_BINDIR ${CMAKE_INSTALL_BINDIR}
+	CACHE PATH "Installation directory for executables")
+set(INSTALL_INCLUDEDIR ${CMAKE_INSTALL_INCLUDEDIR}
+	CACHE PATH "Installation directory for header files")
+if(WIN32 AND NOT CYGWIN)
+	set(DEF_INSTALL_CMAKEDIR CMake)
+else()
+	set(DEF_INSTALL_CMAKEDIR share/cmake/${PROJECT_NAME})
+endif()
+set(INSTALL_CMAKEDIR ${DEF_INSTALL_CMAKEDIR}
+	CACHE PATH "Installation directory for CMake files")
+
+# Report to user
+foreach(p LIB BIN INCLUDE CMAKE)
+	file(TO_NATIVE_PATH ${CMAKE_INSTALL_PREFIX}/${INSTALL_${p}DIR} _path )
+	message(STATUS "Installing ${p} components to ${_path}")
+	unset(_path)
+endforeach()
+
+add_subdirectory(src)
 
 enable_testing()
-include(CTest)
-endif()
 
-set(EXCEPTIONS_SOURCES
-	bug.hpp
-	contract.${VLINDER_TARGET_PLATFORM_NAME}.cpp
-	contract.hpp
-	details/prologue.hpp
-	exception.hpp
-	)
+add_subdirectory(tests)
 
-add_library(exceptions SHARED
-	${EXCEPTIONS_SOURCES}
-	)
-set_target_properties(exceptions PROPERTIES VERSION ${PROJECT_VERSION})
-set_target_properties(exceptions PROPERTIES SOVERSION 1)
-install(TARGETS exceptions
-	EXPORT ExceptionsTargets
-	LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
-	)
-install(FILES bug.hpp contract.hpp exception.hpp DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/exceptions)
-install(FILES details/prologue.hpp DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/exceptions/details)
-
-include(CMakePackageConfigHelpers)
-write_basic_package_version_file(
-	"${CMAKE_CURRENT_BINARY_DIR}/exceptions/exceptions-config-version.cmake"
-	VERSION ${PROJECT_VERSION}
-	COMPATIBILITY SameMajorVersion
-)
-
-export(EXPORT ExceptionsTargets
-	FILE "${CMAKE_CURRENT_BINARY_DIR}/exceptions/exceptions-targets.cmake"
-	NAMESPACE Vlinder::
-)
-configure_file(exceptions-config.cmake "${CMAKE_CURRENT_BINARY_DIR}/exceptions/exceptions-config.cmake"
-	COPYONLY
-)
-
-set(ConfigPackageLocation lib/cmake/exceptions)
-install(EXPORT ExceptionsTargets
-	FILE exceptions-targets.cmake
-	NAMESPACE Vlinder::
-	DESTINATION ${ConfigPackageLocation}
-)
-install(
-	FILES exceptions-config.cmake "${CMAKE_CURRENT_BINARY_DIR}/exceptions/exceptions-config-version.cmake"
-	DESTINATION ${ConfigPackageLocation}
-	COMPONENT Devel
-)
-
-if (${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_SOURCE_DIR})
-	foreach (test
-		exceptions
-		bug
-		)
-		add_executable(test_${test} tests/${test}.cpp)
-		target_link_libraries(test_${test} exceptions ${CMAKE_THREAD_LIBS_INIT})
-		add_test(test_${test} ${EXECUTABLE_OUTPUT_PATH}/test_${test})
-	endforeach (test)
-endif()
-
-# add a target to generate API documentation with Doxygen
-find_package(Doxygen)
-if(DOXYGEN_FOUND)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/Doxyfile.in ${CMAKE_CURRENT_BINARY_DIR}/Doxyfile @ONLY)
-add_custom_target(exceptions-doc
-	${DOXYGEN_EXECUTABLE} ${CMAKE_CURRENT_BINARY_DIR}/Doxyfile
-	WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-	COMMENT "Generating API documentation with Doxygen" VERBATIM
-	)
-endif(DOXYGEN_FOUND)
+include(CMakeCPack.cmake)
