--- CPack.Description.txt
@@ -1,5 +1,0 @@
-DESCRIPTION
-===========
-
-This is Vlinder Software's exceptions library
-
