--- details/prologue.hpp
@@ -1,30 +1,0 @@
-/* Copyright 2016  Vlinder Software
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License. */
-#ifndef exceptions_details_prologue_h
-#define exceptions_details_prologue_h
-
-#ifndef EXCEPTIONS_API
-#	if defined(VLINDER_TARGET_WIN32)
-#		if defined(exceptions_EXPORTS) || defined(exceptions_relacy_EXPORTS)
-#			define EXCEPTIONS_API __declspec(dllexport)
-#		else
-#			define EXCEPTIONS_API __declspec(dllimport)
-#		endif
-#	else
-#		define EXCEPTIONS_API
-#	endif
-#endif
-
-#endif
-
-
