--- exceptions-config.cmake
@@ -1,2 +1,0 @@
-include("${CMAKE_CURRENT_LIST_DIR}/exceptions-targets.cmake")
-set(exceptions_INCLUDE_DIR   "${CMAKE_CURRENT_LIST_DIR}/../../../include/" )
