+++ src/contract.win32.cpp
